Move _HttpClient outside of the whatsapp package
